requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
PyPDF2>=3.0.0
langchain>=0.0.300
langchain-google-genai>=0.0.5
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.21.0
faiss-cpu>=1.7.0
tiktoken>=0.5.0
reportlab>=4.0.0